Los tests pasan correctamente. Creo que ya he finalizado con todo. He tenido que cmabiar un par de cosas, ya he señalado varias cosas a no tener en cuenta, pero estoy demasiado agotado. Maña pongo los comentarios. En teoria ya va. Tambien hay una cosa a aclarar en la base de datos, ya iré informando...

+++ trivia-api-master/src/main/java/com/example/trivia/repository/AnswerRepository.java
@@ -9,5 +9,6 @@ import java.util.List;
 
 @Repository
 public interface AnswerRepository extends CrudRepository<Answer, Long> {
-    List<Answer> findByRoundIdAndQuestionId(Long roundId, Long questionId);
+
+    List<Answer> findByRoundIdAndQuestionId(Integer roundId, Long questionId);
 }
